requirements.txt
Django==1.11.3
django-ckeditor==5.3.0
django-js-asset==1.0.0
django-mptt==0.8.7
django-solo==1.1.2
MySQL-python==1.2.5
olefile==0.44
Pillow==4.2.1
pytz==2017.2
